Moved backpack configs
